# ai/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Fishio API",
    "version": "v1"
  },
  "paths": {
...
    "/api/lookup/category-definitions": {
      "get": {
        "tags": [
          "LookupData"
        ],
        "operationId": "GetCategoryDefinitions",
        "parameters": [
          {
            "name": "FilterByType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CategoryType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDefinitionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/results/{token}": {
      "get": {
        "tags": [
          "PublicResults"
        ],
        "operationId": "GetResultsByToken",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Void"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCalculationLogic": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "CategoryDefinitionDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          },
          "metric": {
            "$ref": "#/components/schemas/CategoryMetric"
          },
          "calculationLogic": {
            "$ref": "#/components/schemas/CategoryCalculationLogic"
          },
          "entityType": {
            "$ref": "#/components/schemas/CategoryEntityType"
          },
          "requiresSpecificFishSpecies": {
            "type": "boolean"
          },
          "isGlobal": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CategoryEntityType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "CategoryMetric": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "CategoryType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Void": {
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid JWT token. Example: \"Bearer {token}\"",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}